added jwt features

+++ resources/item.py
@@ -1,6 +1,12 @@
 
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import (
+	jwt_required,
+	get_jwt_claims,
+	jwt_optional,
+	get_jwt_identity,
+	fresh_jwt_required
+	)
 from models.item import ItemModel 
 
 
@@ -30,7 +36,7 @@ class Item(Resource):
 		return {'message' : 'Item not found'}, 404
 	
 
-	
+	@fresh_jwt_required
 	def post(self, name):
 	
 		if ItemModel.find_by_name(name):
@@ -50,12 +56,15 @@ class Item(Resource):
 				
 	@jwt_required  #need authenticate before calling the jwt method. 	
 	def delete(self, name):
-
+		claims = get_jwt_claims()
+		if not claims['is_admin']:
+			return {'message' : 'Admin priviledge required.'}, 401
 		item = ItemModel.find_by_name(name)
 		if item:
 			item.delete_from_db()
-		return {'message': 'Item deleted'}
-		
+			return {'message': 'Item deleted.'}
+		return {'message': 'Item not found.'}, 404
+
 	def put(self, name):
 
 		data = Item.parser.parse_args()	
@@ -74,7 +83,13 @@ class Item(Resource):
 
 class ItemList(Resource):
 
+	@jwt_optional
 	def get(self):
-	
-		#return {'items' : [item.json for item in ItemModel.query.all()]}
-		return {'items' : list(map(lambda x : x.json(), ItemModel.find_all()))}
+		user_id = get_jwt_identity()
+		items = list(map(lambda x : x.json(), ItemModel.find_all()))
+		if user_id:
+			return {'items' : items, 'user_id' : user_id}, 200
+		return {'items': [item['name'] for item in items],
+				'message' : 'More data available if you log in.',
+				'user_id' : user_id }, 200
+
